core: Remove WigdetPlus in favor of RelmWidgetExt
