+++ examples/relm4-components/file_dialogs.rs
@@ -1,7 +1,7 @@
 use gtk::prelude::*;
 use relm4::{
     gtk, Component, ComponentController, ComponentParts, ComponentSender, Controller, RelmApp,
-    SimpleComponent, WidgetPlus,
+    RelmWidgetExt, SimpleComponent,
 };
 use relm4_components::{open_dialog::*, save_dialog::*};
