--- relm4/src/util/mod.rs
@@ -1,7 +1,0 @@
-//! Utility traits for working with GTK widgets.
-
-mod widget_plus;
-mod widget_ref;
-
-pub use widget_plus::WidgetPlus;
-pub use widget_ref::WidgetRef;
